requirements.txt
pymongo
langchain==0.0.184
python-dotenv==1.0.0
huggingface-hub==0.14.1
google-generativeai